pyproject.toml
[tool.poetry.dependencies]
python = "^3.10"
discord.py = "^2.3.2"
python-dotenv = "^1.0.1"
